gateway/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
redis>=4.5.0
PyJWT==2.8.0
redis[async]
python-jose
prometheus-client
